Tidy screen test
